refactor(frontend): move Protected guard and ConfirmationDialogProvider into rework
Both were the only legacy files reaching back into rework (Protected via
useUserCapabilities, ConfirmationDialogProvider via the ConfirmationDialog
molecule it wraps), which blocked ever treating src/components/ as an
isolated island safe to delete. Protected joins the other route guards in
rework/core/guards/; ConfirmationDialogProvider is co-located with its
ConfirmationDialog molecule, mirroring the existing Toast/ToastProvider
pattern. No behavior change — 9 import sites updated, no logic touched.
