+++ apps/frontend/src/rework/components/shared/molecules/ConfirmationDialog/ConfirmationDialogProvider.tsx
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import { ConfirmationDialog } from "../rework/components/shared/molecules/ConfirmationDialog/ConfirmationDialog";
+import { ConfirmationDialog } from "./ConfirmationDialog";
 import { createContext, PropsWithChildren, useContext, useState } from "react";
 import { useTranslation } from "react-i18next";
